pyproject.toml
[project]
name = "currency-agent"
version = "0.1.0"
description = "Currency conversion agent using ADK, MCP and A2A"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.6",
    "fastmcp>=2.12.3",
    "google-adk>=1.14.1",
    "google-genai>=1.38.0",
    "python-dotenv>=1.1.1",
]
